+++ contracts/test_api/test_string.cpp
@@ -5,7 +5,7 @@
 
 unsigned int test_string::construct_with_size() {
 
-   uint32_t size = 100;
+   size_t size = 100;
    eosio::string str(size);
 
    WASM_ASSERT( str.get_size() == size,  "str.get_size() == size" );
@@ -16,7 +16,7 @@ unsigned int test_string::construct_with_size() {
 
 unsigned int test_string::construct_with_data() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(data, size, false);
 
@@ -29,8 +29,8 @@ unsigned int test_string::construct_with_data() {
 
 unsigned int test_string::construct_with_data_partially() {
   char data[] = "abcdefghij";
-  uint32_t substr_size = 5;
-  uint32_t offset = 2;
+  size_t substr_size = 5;
+  size_t offset = 2;
 
   eosio::string str(data + offset, substr_size, false);
 
@@ -46,7 +46,7 @@ unsigned int test_string::construct_with_data_partially() {
 
 unsigned int test_string::construct_with_data_copied() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(data, size, true);
 
@@ -62,7 +62,7 @@ unsigned int test_string::construct_with_data_copied() {
 
 unsigned int test_string::copy_constructor() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str1(data, size, true);
 
@@ -79,7 +79,7 @@ unsigned int test_string::copy_constructor() {
 
 unsigned int test_string::assignment_operator() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str1(data, size, true);
 
@@ -97,7 +97,7 @@ unsigned int test_string::assignment_operator() {
 
 unsigned int test_string::index_operator() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(data, size, false);
 
@@ -110,7 +110,7 @@ unsigned int test_string::index_operator() {
 
 unsigned int test_string::index_out_of_bound() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(data, size, false);
   char c = str[size];
@@ -120,12 +120,12 @@ unsigned int test_string::index_out_of_bound() {
 
 unsigned int test_string::substring() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(data, size, false);
 
-  uint32_t substr_size = 5;
-  uint32_t offset = 2;
+  size_t substr_size = 5;
+  size_t offset = 2;
   eosio::string substr = str.substr(offset, substr_size, false);
 
   WASM_ASSERT( substr.get_size() == substr_size,  "str.get_size() == substr_size" );
@@ -140,12 +140,12 @@ unsigned int test_string::substring() {
 
 unsigned int test_string::substring_out_of_bound() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(data, size, false);
 
-  uint32_t substr_size = size;
-  uint32_t offset = 1;
+  size_t substr_size = size;
+  size_t offset = 1;
   eosio::string substr = str.substr(offset, substr_size, false);
 
   return WASM_TEST_PASS;
@@ -153,11 +153,12 @@ unsigned int test_string::substring_out_of_bound() {
 
 unsigned int test_string::concatenation_null_terminated() {
   char data1[] = "abcdefghij";
-  uint32_t size1 = sizeof(data1)/sizeof(char);
+
+  size_t size1 = sizeof(data1)/sizeof(char);
   eosio::string str1(data1, size1, false);
 
   char data2[] = "klmnoppqrst";
-  uint32_t size2 = sizeof(data2)/sizeof(char);
+  size_t size2 = sizeof(data2)/sizeof(char);
   eosio::string str2(data2, size2, false);
 
   str1 += str2;
@@ -176,11 +177,12 @@ unsigned int test_string::concatenation_null_terminated() {
 
 unsigned int test_string::concatenation_non_null_terminated() {
   char data1[] = {'a','b','c','d','e','f','g','h','i','j'};
-  uint32_t size1 = sizeof(data1)/sizeof(char);
+
+  size_t size1 = sizeof(data1)/sizeof(char);
   eosio::string str1(data1, size1, false);
 
   char data2[] = {'k','l','m','n','o','p','q','r','s','t'};
-  uint32_t size2 = sizeof(data2)/sizeof(char);
+  size_t size2 = sizeof(data2)/sizeof(char);
   eosio::string str2(data2, size2, false);
 
   str1 += str2;
@@ -199,7 +201,7 @@ unsigned int test_string::concatenation_non_null_terminated() {
 
 unsigned int test_string::assign() {
   char data[] = "abcdefghij";
-  uint32_t size = sizeof(data)/sizeof(char);
+  size_t size = sizeof(data)/sizeof(char);
 
   eosio::string str(100);
   str.assign(data, size, true);
@@ -217,35 +219,35 @@ unsigned int test_string::assign() {
 
 unsigned int test_string::comparison_operator() {
   char data1[] = "abcdefghij";
-  uint32_t size1 = sizeof(data1)/sizeof(char);
+  size_t size1 = sizeof(data1)/sizeof(char);
   eosio::string str1(data1, size1, false);
 
   char data2[] = "abcdefghij";
-  uint32_t size2 = sizeof(data2)/sizeof(char);
+  size_t size2 = sizeof(data2)/sizeof(char);
   eosio::string str2(data2, size2, false);
 
   char data3[] = "klmno";
-  uint32_t size3 = sizeof(data3)/sizeof(char);
+  size_t size3 = sizeof(data3)/sizeof(char);
   eosio::string str3(data3, size3, false);
 
   char data4[] = "aaaaaaaaaaaaaaa";
-  uint32_t size4 = sizeof(data4)/sizeof(char);
+  size_t size4 = sizeof(data4)/sizeof(char);
   eosio::string str4(data4, size4, false);
 
   char data5[] = "你好";
-  uint32_t size5 = sizeof(data5)/sizeof(char);
+  size_t size5 = sizeof(data5)/sizeof(char);
   eosio::string str5(data5, size5, false);
 
   char data6[] = "你好嗎？";
-  uint32_t size6 = sizeof(data6)/sizeof(char);
+  size_t size6 = sizeof(data6)/sizeof(char);
   eosio::string str6(data6, size6, false);
 
   char data7[] = {'a', 'b', 'c', 'd', 'e'};
-  uint32_t size7 = sizeof(data7)/sizeof(char);
+  size_t size7 = sizeof(data7)/sizeof(char);
   eosio::string str7(data7, size7, false);
 
   char data8[] = {'a', 'b', 'c'};
-  uint32_t size8 = sizeof(data8)/sizeof(char);
+  size_t size8 = sizeof(data8)/sizeof(char);
   eosio::string str8(data8, size8, false);
 
   WASM_ASSERT( str1 == str2,  "str1 == str2" );
@@ -263,7 +265,8 @@ unsigned int test_string::comparison_operator() {
 
 unsigned int test_string::print_null_terminated() {
   char data[] = "Hello World!";
-  uint32_t size = sizeof(data)/sizeof(char);
+
+  size_t size = sizeof(data)/sizeof(char);
   eosio::string str(data, size, false);
 
   eosio::print(str);
@@ -273,7 +276,8 @@ unsigned int test_string::print_null_terminated() {
 
 unsigned int test_string::print_non_null_terminated() {
   char data[] = {'H', 'e', 'l', 'l', 'o', ' ', 'W', 'o', 'r', 'l', 'd', '!'};
-  uint32_t size = sizeof(data)/sizeof(char);
+
+  size_t size = sizeof(data)/sizeof(char);
   eosio::string str(data, size, false);
 
   eosio::print(str);
@@ -283,7 +287,8 @@ unsigned int test_string::print_non_null_terminated() {
 
 unsigned int test_string::print_unicode() {
   char data[] = "你好，世界！";
-  uint32_t size = sizeof(data)/sizeof(char);
+
+  size_t size = sizeof(data)/sizeof(char);
   eosio::string str(data, size, false);
 
   eosio::print(str);
